refactor: split request tasks
